git-svn-id: https://5-10/svn/u309m_common@4 784819df-0e49-ba48-a535-3ffbe095aa23

+++ common_data.h
@@ -520,7 +520,7 @@ struct control_data_t
   irs::bit_data_t izm_th_spi_enable;
   
   irs::conn_data_t<irs_u32> connect_counter;
-  irs::conn_data_t<double> time;
+  irs::conn_data_t<long> time;
 
   control_data_t(irs::mxdata_t *ap_data = IRS_NULL, irs_uarc a_index = 0,
     irs_uarc* ap_size = IRS_NULL)
@@ -617,7 +617,7 @@ struct eth_data_t {
   supply_eth_data_t supply_2V;    //  114 bytes
   supply_eth_data_t supply_1A;    //  114 bytes
   supply_eth_data_t supply_17A;   //  114 bytes
-  control_data_t control;         //  20 bytes
+  control_data_t control;         //  18 bytes
   //---------------------------------------------
   //                          Итого:  668 байт
